Fix gtimer on freertos
gtime tries to calculate from ticks to milliseconds by dividing by
`gfxMillisecondsToTicks(1)`. That'd be fine on systems where there's multiple
milliseconds per tick, but on freertos we count 10 ticks per millisecond, so
`gfxMillisecondsToTicks(1)` is 0, causing a division by zero later on.

For a 'proper' fix we'd need to implement gfxTicksToMilliseconds
on each backend, not just freertos