Don't detach the toggle source on widget destruction
AFAICS this source is being reused when creating other widgets later, so
if we detach it when destroying the first widget, a subsequent widget instance
will not receive events.

I encountered this problem when creating a List, then destroying that List,
then creating a new List. The new List did not receive any toggle events.